requirements.txt
# --- Core Logic & Data ---
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.5.0
joblib==1.3.2
python-dotenv==1.0.1

# --- API & Server ---
flask==3.0.2
flask-cors==4.0.1

# --- AI & LLM ---
# REQUIRED: The new V1 SDK used by your gen_ai_module.py
google-genai
# REQUIRED: The old SDK used by your report_generator.py
google-generativeai
openai==1.36.0

# --- PDF Generation ---
reportlab==4.0.4

# --- Visualization & Extras (Keep if needed for analysis) ---
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.21.0
pillow==10.2.0
scipy==1.13.0
xgboost==2.0.3
imbalanced-learn==0.12.3
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
tensorflow==2.15.0